requirements.txt
beautifulsoup4
certifi
fabric
httplib2
python-dateutil
pytz
pyyaml
requests
slugify
tinycss
uritemplate
